+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Bug fixes:
 Compatibility:
 
 * Make interpolated strings frozen for compatibility with Ruby 2.7 (#2304, @kirs).
+* Include the type name in exception messages from `rb_check_type` (#2307).
 
 Performance: